% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_DFS2020.R
\docType{data}
\name{data.DFS2020}
\alias{data.DFS2020}
\title{Data of Dvorak, Fischbacher and Schmelz (2020)}
\format{
A data frame with 569 rows and 8 variables:
\describe{
  \item{id}{A vector of integers which identifies the participant.}
  \item{game}{A vector of integers which identifies the game.}
  \item{period}{A vector of integers which identifies the period.}
  \item{others.choices}{A factor with two levels which indicates if the choices of the other two group members are in line (in line) or not in line (not in line) with the preference of the participant.}
  \item{choice}{A factor with two levels which is indicates if the participant deviates (deviate) from or follows (follow) the own preference in the current period.}
  \item{intercept}{An intercept. One for every observation.}
  \item{conformity.score}{The conformity score of the participant in a post-experimental conformity questionnaire (Mehrabian and Stefl, 1995).}
  \item{input}{A factor which indicates the input for the strategies. The variable choice with two levels which is indicates if the participant deviates (deviate) from or follows (follow) the own preference in the current period.}
}
}
\usage{
data(data.DF2011)
}
\description{
A stratEst.data object with observations from an experiment on social influence.
}
\references{
Dvorak F, Fischbacher U, Schmelz K (2020). "Incentives for Conformity and Anticonformity." \emph{TWI Working Paper Series}.

Mehrabian A, Stefl CA (1995). "Basic Temperament Components of Loneliness, Shyness, and Conformity." \emph{Social Behavior and Personality}, 23(3), 253–263.
}
\keyword{datasets}
