% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DF2011all.R
\docType{data}
\name{DF2011all}
\alias{DF2011all}
\title{Dal Bo and Frechette data.}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{coop}{cooperation}
  \item{date}{date of the session}
  \item{delta}{discount factor of treatment}
  \item{group}{group id of matches}
  \item{id}{id of participant}
  \item{period}{period of the supergame}
  \item{r}{stage game parameter of treatment}
  \item{supergame}{supergame number}
}}
\source{
\url{https://www.aeaweb.org/articles?id=10.1257/aer.101.1.411}
}
\usage{
DF2011all
}
\description{
A dataset containing observations of the repeated prisonner's dilemma. zip file dal_bo_2010a_data.txt
}
\keyword{datasets}
