% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTF2T.R
\docType{data}
\name{DTF2T}
\alias{DTF2T}
\title{DTF2T}
\format{A data frame with 8 variables:
\describe{
  \item{\code{state}}{Indicates the state of the automaton.}
  \item{\code{r0}}{Probability to play D (output is 0).}
  \item{\code{r1}}{Probability to play C (output is 1).}
  \item{\code{tremble}}{Probability of a tremble.}
  \item{\code{t1}}{State transition if the action profile is CC (input is 1).}
  \item{\code{t2}}{State transition if the action profile is CD (input is 2).}
  \item{\code{t3}}{State transition if the action profile is DC (input is 3).}
  \item{\code{t4}}{State transition if the action profile is DD (input is 4).}
}}
\usage{
data(DTF2T)
}
\description{
Matrix representation of the prisoner's dilemma strategy which plays D in the first round, then play TF2T.
}
\examples{
strategies <- rbind(DTF2T,GRIM,ALLD,ALLC,TFT)
}
\keyword{datasets}
