% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_stratEst_check.R
\name{is.stratEst.check}
\alias{is.stratEst.check}
\title{Class stratEst.check}
\usage{
is.stratEst.check(x)
}
\arguments{
\item{x}{object to be tested.}
}
\value{
\code{is.stratEst.check} returns \code{TRUE} if its argument is a \code{stratEst.check} object (that is, has "stratEst.check" among its classes) and \code{FALSE} otherwise.
}
\description{
Checks if an object is of class \code{stratEst.check}.
}
\details{
Objects of class \code{stratEst.check} are returned by the function \code{stratEst.check()} of package \code{stratEst}.
}
