% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_stratEst_model.R
\name{is.stratEst.model}
\alias{is.stratEst.model}
\title{Class stratEst.model}
\usage{
is.stratEst.model(x)
}
\arguments{
\item{x}{object to be tested.}
}
\description{
Checks if an object is of class \code{stratEst.model}.
}
\details{
Objects of class \code{stratEst.model} are returned by the estimation function \code{stratEst.model()} of package \code{stratEst}.
}
