% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_stratEst.R
\name{is.stratEst}
\alias{is.stratEst}
\title{Class stratEst}
\usage{
is.stratEst(x)
}
\arguments{
\item{x}{object to be tested.}
}
\description{
Checks if an object is of class \code{stratEst}.
}
\details{
Objects of class \code{stratEst} are returned by the estimation function \code{stratEst()} of package \code{stratEst}.
}
