% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_stratEst_strategy.R
\name{is.stratEst.strategy}
\alias{is.stratEst.strategy}
\title{Class stratEst.strategy}
\usage{
is.stratEst.strategy(x)
}
\arguments{
\item{x}{object to be tested.}
}
\description{
Checks if an object is of class \code{stratEst.strategy}.
}
\details{
Objects of class \code{stratEst.strategy} is returned by the function \code{stratEst.strategy()} of package \code{stratEst}.
}
