% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M1BF.R
\docType{data}
\name{M1BF}
\alias{M1BF}
\title{M1BF}
\format{A matrix with 1 row and 6 columns:
\describe{
  \item{Rows}{Each row corresponds to one states of the automaton.}{}
  \item{Column 1}{Enumerates the states of the automaton.}
  \item{Column 2}{Probability to play C given the current state of the automaton. }
  \item{Column 3}{State transition if the history of play in the last round was CC (input is 1).}
  \item{Column 4}{State transition if the history of play in the last round was CD (input is 2).}
  \item{Column 5}{State transition if the history of play in the last round was DC (input is 3).}
  \item{Column 6}{State transition if the history of play in the last round was DD (input is 4).}
}}
\usage{
data(M1BF)
}
\description{
Matrix represention of the prisoner's dilemma strategy which plays if both players played C, and D if both players played D. If the own action was C and the other player played D, play C with some probability. If the own action was D and the other player played C, play C with some (potentially different) probability.
}
\examples{
strategies <- rbind(M1BF,GRIM,ALLD,ALLC,TFT)
}
\keyword{datasets}
