% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stratEst_strategy.R
\name{plot.stratEst.strategy}
\alias{plot.stratEst.strategy}
\title{Plot Method for stratEst.strategy}
\usage{
\method{plot}{stratEst.strategy}(
  x,
  y,
  ...,
  title = NULL,
  show.legend = T,
  show.title = T,
  node.fontsize = 25,
  main.fontsize = 40,
  arrow.fontsize = 20,
  legend.fontsize = 20,
  legend.width = 1,
  node.width = 1,
  arrowsize = 1,
  node.penwidth = 1,
  arrow.penwidth = 1,
  fillcolor = NULL,
  ranksep = 0,
  file = NA
)
}
\arguments{
\item{x}{An object of class \code{stratEst.strategy}.}

\item{y}{Argument two of the generic function.}

\item{...}{Further arguments passed to or from other methods.}

\item{title}{String. The title of the plot.}

\item{show.legend}{Logical. Hide plot legend if FALSE. Default is TRUE.}

\item{show.title}{Logical. Hide plot title if FALSE. Default is TRUE.}

\item{node.fontsize}{Font-size of the plot labels.}

\item{main.fontsize}{Font-size of the plot title.}

\item{arrow.fontsize}{Font-size of the arrow labels.}

\item{legend.fontsize}{Font-size of the legend.}

\item{legend.width}{Width of the legend items.}

\item{node.width}{Width of the nodes.}

\item{arrowsize}{Size of the arrowhead.}

\item{node.penwidth}{Width of the nodes.}

\item{arrow.penwidth}{Width of the nodes.}

\item{fillcolor}{Vector of hex-color codes of the choices.}

\item{ranksep}{Separation of nodes with the same rank.}

\item{file}{String. A valid path followed by a file name. Should end with .svg. Default is NA and no file is written.}
}
\description{
Plot Method for stratEst.strategy
}
