% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_stratEst_check.R
\name{print.stratEst.check}
\alias{print.stratEst.check}
\title{Print Method for stratEst.check}
\usage{
\method{print}{stratEst.check}(x, ...)
}
\arguments{
\item{x}{An object of class \code{stratEst.check}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print Method for stratEst.check
}
