% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_stratEst_data.R
\name{print.stratEst.data}
\alias{print.stratEst.data}
\title{Print Method for stratEst.data}
\usage{
\method{print}{stratEst.data}(x, ...)
}
\arguments{
\item{x}{An object of class \code{stratEst.data}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print Method for stratEst.data
}
