% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_stratEst_model.R
\name{print.stratEst.model}
\alias{print.stratEst.model}
\title{Print Method for stratEst.model}
\usage{
\method{print}{stratEst.model}(x, ...)
}
\arguments{
\item{x}{An object of class \code{stratEst.model}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print Method for stratEst.model
}
