% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_stratEst.R
\name{print.stratEst}
\alias{print.stratEst}
\title{Print Method for stratEst}
\usage{
\method{print}{stratEst}(x, ...)
}
\arguments{
\item{x}{An object of class \code{stratEst}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print Method for stratEst
}
