% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_stratEst_strategy.R
\name{print.stratEst.strategy}
\alias{print.stratEst.strategy}
\title{Print Method for stratEst.strategy}
\usage{
\method{print}{stratEst.strategy}(x, ...)
}
\arguments{
\item{x}{An object of class \code{stratEst.strategy}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print Method for stratEst.strategy
}
