% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_stratEst_strategy.R
\name{round.stratEst.strategy}
\alias{round.stratEst.strategy}
\title{Round Method for stratEst.strategy}
\usage{
\method{round}{stratEst.strategy}(x, digits = 0)
}
\arguments{
\item{x}{An object of class \code{stratEst.strategy}.}

\item{digits}{Further arguments passed to or from other methods.}
}
\description{
Round Method for stratEst.strategy
}
