% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGRIM.R
\docType{data}
\name{SGRIM}
\alias{SGRIM}
\title{SGRIM}
\format{A data frame with 8 variables:
\describe{
  \item{\code{state}}{Indicates the state of the automaton.}
  \item{\code{r0}}{Probability to play D (output is 0).}
  \item{\code{r1}}{Probability to play C (output is 1).}
  \item{\code{tremble}}{Probability of a tremble.}
  \item{\code{t1}}{State transition if the action profile is CC (input is 1).}
  \item{\code{t2}}{State transition if the action profile is CD (input is 2).}
  \item{\code{t3}}{State transition if the action profile is DC (input is 3).}
  \item{\code{t4}}{State transition if the action profile is DD (input is 4).}
}}
\usage{
data(SGRIM)
}
\description{
Matrix representation of the semi grim strategy (Breitmoser 2015). The strategy plays C if both players played C, and D if both players played D. If one player played D and the other C, play C with some probability.
}
\examples{
strategies <- rbind(SGRIM,GRIM,ALLD,ALLC,TFT)
}
\references{
Breitmoser, Y. (2015): Cooperation, but no reciprocity: Individual strategies in the repeated prisoner's dilemma, \emph{American Economic Review}, 105, 2882-2910.
}
\keyword{datasets}
