% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stratEst_cpp}
\alias{stratEst_cpp}
\title{The stratEst function}
\usage{
stratEst_cpp(data, strategies, shares, covariates, cluster,
  response = "mixed", r_responses = "no", r_trembles = "global",
  select = "no", crit = "BIC", SE = "yes", outer_runs = 10L,
  outer_tol_eval = 0, outer_max_eval = 1000L, inner_runs = 100L,
  inner_tol_eval = 0, inner_max_eval = 100L, LCR_runs = 100L,
  LCR_tol_eval = 0L, LCR_max_eval = 1000L, BS_samples = 1000L,
  newton_stepsize = 1L, penalty = 0)
}
\arguments{
\item{data}{The data matrix.}
}
\description{
The stratEst function
}
