% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stratEst_check.R
\name{summary.stratEst.check}
\alias{summary.stratEst.check}
\title{Method dispatch for Generic Function Summary}
\usage{
\method{summary}{stratEst.check}(object, ...)
}
\arguments{
\item{object}{An object returned by the function\code{stratEst.check()}. An object of class \code{stratEst.check}.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Method dispatch for Generic Function Summary
}
