% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stratEst_data.R
\name{summary.stratEst.data}
\alias{summary.stratEst.data}
\title{Method dispatch for Generic Function summary}
\usage{
\method{summary}{stratEst.data}(object, ...)
}
\arguments{
\item{object}{An object to be summarized.}

\item{...}{additional arguments affecting the result.}
}
\description{
Method dispatch for Generic Function summary
}
