% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stratEst_model.R
\name{summary.stratEst.model}
\alias{summary.stratEst.model}
\title{Dethod dispatch for Generic Function Summary}
\usage{
\method{summary}{stratEst.model}(object, ...)
}
\arguments{
\item{object}{An object returned by the estimation function\code{stratEst.model()}. An object of class \code{stratEst.model}.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Dethod dispatch for Generic Function Summary
}
