% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stratEst.R
\name{summary.stratEst}
\alias{summary.stratEst}
\title{Dethod dispatch for Generic Function Summary}
\usage{
\method{summary}{stratEst}(object, ...)
}
\arguments{
\item{object}{An object returned by the estimation function\code{stratEst()}. An object of class \code{stratEst}.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Dethod dispatch for Generic Function Summary
}
