% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DF2011.R
\docType{data}
\name{DF2011}
\alias{DF2011}
\title{Data of Dal Bo and Frechette (2011)}
\format{
A data frame with 7358 rows and 6 variables:
\describe{
  \item{treatment}{A factor with six levels which identifies the treatments of the experiment.}
  \item{id}{A vector of integers which identifies the participant.}
  \item{game}{A vector of integers which identifies the supergame.}
  \item{period}{A vector of integers which identifies the period of the supergame.}
  \item{choice}{A factor with two levels which is indicates if the participant cooperates (c) or defects (d) in the current period.}
  \item{other_choice}{A factor with two levels which indicates if the other participant cooperates (c) or defects (d) in the current period.}
}
}
\source{
\url{https://www.aeaweb.org/articles?id=10.1257/aer.101.1.411}
}
\usage{
data(DF2011)
}
\description{
A dataset with observations from the repeated prisoner's dilemma experiment of Dal Bo and Frechette (2011).
}
\references{
Dal Bo, P. and G. R. Frechette (2011): The evolution of cooperation in infinitely repeated games: Experimental evidence, \emph{American Economic Review}, 101, 411-429.
}
\keyword{datasets}
