% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DF2011.R
\docType{data}
\name{DF2011}
\alias{DF2011}
\title{Dal Bo and Frechette data.}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{http://www.aer.info/}
}
\usage{
DF2011
}
\description{
A dataset containing observations of the repeated prisonner's dilemma.
}
\keyword{datasets}
