% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_stratEst_data.R
\name{is.stratEst.data}
\alias{is.stratEst.data}
\title{Class stratEst.data}
\usage{
is.stratEst.data(x)
}
\arguments{
\item{x}{object to be tested.}
}
\value{
\code{is.stratEst.data} returns \code{TRUE} if its argument is a \code{stratEst.data} object (that is, has "stratEst.data" amongst its classes) and \code{FALSE} otherwise.
}
\description{
Checks if an object is of class \code{stratEst.data}.
}
\details{
Objects of class \code{stratEst.data} are returned by the functions \code{stratEst.data()} and \code{stratEst.simulate()} of package \code{stratEst}.
}
