% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_stratEst_model.R
\name{is.stratEst.model}
\alias{is.stratEst.model}
\title{Class stratEst.model}
\usage{
is.stratEst.model(x)
}
\arguments{
\item{x}{object to be tested.}
}
\value{
\code{is.stratEst.model} returns \code{TRUE} if its argument is a \code{stratEst.model} object (that is, has "stratEst.model" amongst its classes) and \code{FALSE} otherwise.
}
\description{
Checks if an object is of class \code{stratEst.model}.
}
\details{
Objects of class \code{stratEst.model} are returned by the estimation function \code{stratEst.model()} of package \code{stratEst}.
}
