% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_stratEst_strategy.R
\name{is.stratEst.strategy}
\alias{is.stratEst.strategy}
\title{Class stratEst.strategy}
\usage{
is.stratEst.strategy(x)
}
\arguments{
\item{x}{object to be tested.}
}
\value{
\code{is.stratEst.strategy} returns \code{TRUE} if its argument is a \code{stratEst.strategy} object (that is, has "stratEst.strategy" amongst its classes) and \code{FALSE} otherwise.
}
\description{
Checks if an object is of class \code{stratEst.strategy}.
}
\details{
Objects of class \code{stratEst.strategy} is returned by the function \code{stratEst.strategy()} of package \code{stratEst}.
}
