% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PT2FT.R
\docType{data}
\name{PT2FT}
\alias{PT2FT}
\title{PT2FT}
\format{A matrix with 1 row and 6 columns:
\describe{
  \item{Rows}{Each row corresponds to one states of the automaton.}{}
  \item{Column 1}{Enumerates the states of the automaton.}
  \item{Column 2}{Probability to play C given the current state of the automaton. }
  \item{Column 3}{State transition if the history of play in the last round was CC (input is 1).}
  \item{Column 4}{State transition if the history of play in the last round was CD (input is 2).}
  \item{Column 5}{State transition if the history of play in the last round was DC (input is 3).}
  \item{Column 6}{State transition if the history of play in the last round was DD (input is 4).}
}}
\usage{
data(PT2FT)
}
\description{
Matrix represention of the prisoner's dilemma strategy which plays C if both players played C in the last 2 rounds, both players played D in the last 2 rounds, or both players played D 2 rounds ago and C last round. Otherwise play D.
}
\examples{
strategies <- rbind(PT2FT,GRIM,ALLD,ALLC,TFT)
}
\keyword{datasets}
