% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratEst_data.R
\name{stratEst.data}
\alias{stratEst.data}
\title{Creates a stratEst.data object.}
\usage{
stratEst.data(
  data,
  choice = "choice",
  id = "id",
  input = "input",
  input.lag = 0,
  input.sep = "",
  game = "game",
  period = "period",
  add = NULL,
  drop = NULL
)
}
\arguments{
\item{data}{A data.frame object in the long format. The object \code{data} is mandatory.}

\item{choice}{A character indicating the variable in \code{data} which contains the discete choices. The default is "choice".}

\item{id}{A character indicating the variable in \code{data} which identifies observations of the same individual across the rows of the data frame. The default is "id".}

\item{input}{A character vector indicating the names of the input generating variables in \code{data}. At least one input generating variable has to be specified.}

\item{input.lag}{A numeric vector indicating how many periods later each variable will affect the generation of the input. The vector must have as many elements as variables specified in the object \code{input}. The object \code{input.lag} is optional. The default is a vector of zeros.}

\item{input.sep}{An optional character used to seperate the values of the input generating variables. Default is no seperation character.}

\item{game}{A character indicating the variable in \code{data} which identifies observations of the same individual in one game across the rows of the data frame. The default is "game".}

\item{period}{A character indicating the variable in \code{data} which identifies the period within a game. The object \code{period} is optional. The default is "period".}

\item{add}{An optional character vector indicating the names of variables in the global environment that should be added to the data frame.}

\item{drop}{An optional character vector indicating the names of variables in \code{data} that should be droped from the data frame.}
}
\value{
A stratEst.data object, which is a data.frame in the long format with the following variables:
\item{data}{A data.frame object in the format required by the estimation function.}
}
\description{
Creates a stratEst.data object.
}
\details{
The data generation function \code{stratEst.data()} transforms data into the format required by the estimation function \code{stratEst()}.
}
\references{
Dal Bo, P. and G. R. Frechette (2011): The evolution of cooperation in infinitely repeated games: Experimental evidence, \emph{American Economic Review}, 101, 411-429.

Fudenberg, D., Rand, D. G., and Dreber, A. (2012): Slow to Anger and Fast to Forgive: Cooperation in an Uncertain World, \emph{American Economic Review}, 102, 720-749.

Wang, Z., Xu, B., and Zhou, H. (2014): Social cycling and conditional responses in the Rock-Paper-Scissors game. \emph{Scientific Reports} 4, 5830.

#' @examples
## Transform the rock-paper-scissors data of Wang, Xu, and Zhou (2014)
data.WXZ2014 <- stratEst.data(WXZ2014, input = c("choice"), choice = "choice", input.lag = 1)

## Transform the prisoner's dilemma data of Dal Bo and Frechette (2011).
data.DF2011 <- stratEst.data(DF2011, choice ="choice", input =c("choice","other.choice"), input.lag = 1)

#' ## Transform the prisoner's dilemma data of Fudenberg, Rand, and Dreber (2012).
data.FRD2012 <- stratEst.data(data = FRD2012, choice ="choice", input =c("last.choice","last.other"))
}
