% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratEst_data.R
\name{stratEst.data}
\alias{stratEst.data}
\title{Creates a stratEst.data object.}
\usage{
stratEst.data(
  data,
  id = "id",
  output,
  input.vars,
  input.lag = 0,
  input.sep = "",
  game = "game",
  period = "period",
  add,
  drop
)
}
\arguments{
\item{data}{A data.frame object in the long format. The object \code{data} is mandatory.}

\item{id}{A character indicating the variable in \code{data} which identifies observations of the same individual across the rows of the data frame. The object \code{id} is mandatory.}

\item{output}{A character indicating the variable in \code{data} which contains the discete choices. The object \code{output} is mandatory.}

\item{input.vars}{A character vector indicating the names of the input generating variables in \code{data}. The object \code{input.vars} is mandatory. At least one input generating variable has to be specified.}

\item{input.lag}{A numeric vector indicating how many periods later each variable will affect the generation of the input. The vector must have as many elements as variables specified in the input object \code{input.vars}. The object \code{input.lag} is optional. The default is a vector of zeros.}

\item{input.sep}{An optional character used to seperate the values of the input generating variables. Default is no seperation character.}

\item{game}{A character indicating the variable in \code{data} which identifies observations of the same individual in one game across the rows of the data frame. . The object \code{game} is optional. The default is a vector of ones.}

\item{period}{A character indicating the variable in \code{data} which identifies the period within a game. The object \code{period} is optional. The default is a vector of ones.}

\item{add}{An optional character vector indicating the names of variables in the global environment that should be added to the data frame.}

\item{drop}{An optional character vector indicating the names of variables in \code{data} that should be droped from the data frame.}
}
\value{
A stratEst.data object, which is a data.frame in the long format with the following variables:
\item{data}{A data.frame object in the format required by the estimation function.}
}
\description{
Creates a stratEst.data object.
}
\details{
The data generation function \code{stratEst.data()} transforms data into the format required by the estimation function \code{stratEst()}.
}
\references{
Dal Bo, P. and G. R. Frechette (2011): The evolution of cooperation in infinitely repeated games: Experimental evidence, \emph{American Economic Review}, 101, 411-429.
#' @examples
## Transform data of the rock-paper-scissors (from https://roshambo.me/) to estimate strategies which react to the previous move in the same game.
data.RPS <- stratEst.data( data = RPS,input.vars = c("choice"), output = "choice", input.lag = 1 )

## Transform the dataset of Dal Bo and Frechette (2011).
data <- stratEst.data(DF2011,input.vars =c("choice","other_choice"),output="choice", input.lag = 1)
}
