% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratEst.R
\name{stratEst}
\alias{stratEst}
\title{A function}
\usage{
stratEst(data, strategies, shares, covariates, cluster, response = "mixed",
  r.responses = "no", r.trembles = "global", select = "no",
  crit = "BIC", se = "yes", outer.runs = 10, outer.tol = 0,
  outer.max = 1000, inner.runs = 100, inner.tol = 0, inner.max = 10,
  lcr.runs = 1000, lcr.tol = 0, lcr.max = 1000, bs.samples = 1000)
}
\arguments{
\item{data}{A datat matrix.}

\item{strategies}{A matrix identifying the strategies.}

\item{shares}{A vector of shares}

\item{covariates}{A vector}

\item{cluster}{A vector}

\item{response}{A vector}

\item{r.responses}{A vector}

\item{r.trembles}{A vector}

\item{select}{A vector}

\item{crit}{A vector}

\item{se}{A vector}

\item{outer.runs}{A number}

\item{outer.tol}{A number}

\item{outer.max}{A vector}

\item{inner.runs}{A vector}

\item{inner.tol}{A vector}

\item{inner.max}{A vector}

\item{lcr.runs}{A vector}

\item{lcr.tol}{A vector}

\item{lcr.max}{A vector}

\item{bs.samples}{A vector}
}
\description{
A function
}
