% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratEst_simulate.R
\name{stratEst.simulate}
\alias{stratEst.simulate}
\title{Simulation function for strategy estimation.}
\usage{
stratEst.simulate(
  data = NULL,
  strategies,
  shares = NULL,
  coefficients = NULL,
  covariate.mat = NULL,
  num.ids = 100,
  num.games = 5,
  num.periods = NULL,
  fixed.assignment = TRUE,
  input.na = FALSE,
  sample.id = NULL
)
}
\arguments{
\item{data}{A \code{stratEst.data} object. Alternatively, the arguments \code{num.ids}, \code{num.games}, and \code{num.periods} can be used if no data is available.}

\item{strategies}{A list of strategies. Each strategy is a data.frame of class \code{stratEst.strategy}. Each row of the data.frame represents one state of the strategy. The first row defines the initial state which is entered if the variable input is NA. Column names which start with the string 'output.' indicate the columns which contain the multinomial choice probabilities of the strategy. For example, a column labeled 'output.x' contains the probability to observe the output 'x'. The column 'tremble' contains a tremble probability for pure strategies. Column names which start with the string 'input.' indicate the columns which contain the deterministic state transition of the strategy. For example, a column with name 'input.x' indicates the state transition after observing input 'x'.}

\item{shares}{A vector of strategy shares. The elements to the order of strategies in the list \code{strategies}. Shares which are \code{NA} are estimated from the data. With more than one sample and sample specific shares, a list of column vectors is required.}

\item{coefficients}{A matrix of regression coefficients. The column names correspond to the names of the strategies. The row names to the names of the covariates.}

\item{covariate.mat}{A matrix with the covariates in columns. The covariate matrix must have as many rows as there are individuals.}

\item{num.ids}{Integer which specifies the number of individuals which are assigned to the stategies. Default is 100.}

\item{num.games}{Integer which specifies the number of games each individual plays. Default is 5 games.}

\item{num.periods}{A column vector with as many elments as there are games. The elements of the vector are integer vaules which specify the number of periods of each game. Default is 5 periods in each game.}

\item{fixed.assignment}{A logical indicating whether the assignment of individuals to strategies is fixed. If \code{FALSE} individuals are repeatedly assigned to strategies for each each game. If \code{TRUE}, individuals are only assigned once, and use the assigned strategy in each game. Default is \code{FALSE}.}

\item{input.na}{A logical indicating whether the input in the first period of a game is \code{NA}. If \code{FALSE}, and input is randomly selected for the first period. Default is \code{FALSE}.}

\item{sample.id}{A character indicating the name of the variable which identifies the samples in data. Individual observations must be nested in samples. The same must be true for clusters if specified. If more than one sample exists, shares are estimated for each sample. All other parameters are estimated for the data of all samples. If the object is not supplied, it is assumed that the data contains only one sample.}
}
\value{
A \code{stratEst.data} object. A data frame in the long format with the following variables:
\item{id}{the individual.}
\item{game}{the game.}
\item{period}{the period.}
\item{choice}{the choice.}
\item{input}{the input.}
}
\description{
Simulates data which can be used to test the strategy estimation function \code{stratEst()}.
}
