% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratEst_strategy.R
\name{stratEst.strategy}
\alias{stratEst.strategy}
\title{Creates a stratEst.strategy object.}
\usage{
stratEst.strategy(
  inputs,
  outputs,
  num.states,
  responses,
  trembles,
  transitions
)
}
\arguments{
\item{inputs}{The names of the inputs, i.e. the levels of factor input in data.}

\item{outputs}{The names of the outputs, i.e. the levels of factor output in data. .}

\item{num.states}{An integer indicating the number states of the strategy.}

\item{responses}{A numeric vector which contains the response probabilities of the strategy.}

\item{trembles}{A numeric vector which contains the tremble probabilities of the strategy.}

\item{transitions}{A vector of integers which contains the deterministic state transitions.}
}
\value{
A stratEst.strategy object. A data.frame with the following variables:
\item{output.x}{The probability for output 'x'.}
\item{trembles}{The probability to observe a tremble.}
\item{input.x}{The deterministic state transition of the strategy after observing of input 'x'.}
}
\description{
Creates a stratEst.strategy object.
}
\details{
The strategy generation function \code{stratEst.strategy()} creates a strategy as a data.frame.
}
\examples{
## Create the mixed nash strategy for the rock-paper-scissors game.
ins = c(NA,"rock","paper","scissors")
outs = c("rock","paper","scissors")
nash =  stratEst.strategy( inputs = ins , outputs = outs , num.states = 1 , responses = 1/3 )
}
