% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratEst_strategy.R
\name{stratEst.strategy}
\alias{stratEst.strategy}
\title{Creates a stratEst.strategy object.}
\usage{
stratEst.strategy(
  choices,
  inputs = NULL,
  prob.choices = NULL,
  tr.inputs = NULL,
  trembles = NULL,
  num.states = NULL
)
}
\arguments{
\item{choices}{A character vector indicating the levels of the variable 'output' in data.}

\item{inputs}{A character vector indicating the levels of the variable 'input' in data.}

\item{prob.choices}{A numeric vector which contains the response probabilities of the strategy.}

\item{tr.inputs}{A vector of integers which contains the deterministic state transitions.}

\item{trembles}{A numeric vector which contains the tremble probabilities of the strategy.}

\item{num.states}{An integer indicating the number states of the strategy.}
}
\value{
A stratEst.strategy object. A data.frame with the following variables:
\item{prob.x}{The probability of choice 'x'.}
\item{tremble}{The probability to observe a tremble.}
\item{tr(x)}{The deterministic state transition of the strategy after input 'x'.}
}
\description{
Creates a stratEst.strategy object.
}
\details{
The strategy generation function \code{stratEst.strategy()} creates a strategy as a data.frame.
}
\examples{
## Nash equilibrium strategy of rock-paper-scissors
ins = c(NA,"rock","paper","scissors")
rps = c("rock","paper","scissors")
mixed = stratEst.strategy(choices = rps)
nash = stratEst.strategy(choices = rps, prob.choices = rep(1/3,3))
rock = stratEst.strategy(choices = rps, prob.choices = c(1,0,0))
}
