% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratEst_test.R
\name{stratEst.test}
\alias{stratEst.test}
\title{Runs t-tests if model parameters differ from user defined values}
\usage{
stratEst.test(
  model,
  par = c("shares", "probs", "trembles", "coefficients"),
  values = 0,
  alternative = "two.sided",
  digits = 4
)
}
\arguments{
\item{model}{A fitted model returned by the estimation function\code{stratEst()}. An object of class \code{stratEst}.}

\item{par}{A character vector indicating the type parameters to be tested. The default is a vector with all possible elements: "shares", "probs", "trembles", and "coefficients".}

\item{values}{A vector of numeric values the parameter are compared to. Default is zero.}

\item{alternative}{A character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "smaller". You can specify just the initial letter.}

\item{digits}{An integer which specifies the number of digits of the results.}
}
\value{
A data.frame with one row for each estimated parameter and 5 columns:
\item{estimate}{The parameter estimate.}
\item{diff}{The difference between the estimated parameter and the numeric value (if values unequal to zero are supplied).}
\item{std.error}{The standard error of the parameter estimate.}
\item{t-value}{The t-statistic.}
\item{res.degrees}{The residual degrees of freedom.}
\item{Pr(>|t|)}{The p-value of the test.}
}
\description{
Runs t-tests if model parameters differ from user defined values
}
\details{
The test function retrieves the parameter estimates and standard errors from the model and performs a two-sided t-test.
}
\examples{
## Nash equilibrium strategy of rock-paper-scissors

}
